chore(nuget): use custom nuget publish (#622)
Since the new Ubuntu build image on DevOps does not have `mono`
installed, it means that some tools like **NuGet** are not built-in
installed anymore. As a migration, they recommend doing it yourself.

This PR uses the `dotnet nuget push` instead (like we did with the MyGet
release), which fixes this problem.