feat(otel): use new observability api for otel/serilog implementations (#618)
Use a new interface called `IObservability` to place traces/metrics away
from the `ILogger` interface. This is the result of the discussion #587
. This PR provides an implementation for both OpenTelemetry and Serilog
to showcase in a single go the possible abstraction and to ease the
migration from Serilog ➡️ OpenTelemetry.

No feature documentation updates are made yet, as this will require a
refactoring exercise by itself.